Require same major version of neutron
Release cycles for the -aas project are in sync with neutron, as well as
major version numbers. Since the -aas code directly imports from neutron
module, we should make sure that the major versions are identical,
otherwise we can get into a situation where e.g. fwaas code is a lot
older than neutron and is not compatible with it, breaking the plugin,
as well as other tools that happen to import fwaas code during their
operations (f.e. neutron-db-manage does it).

Change-Id: I01e97ae3a6db40fefdbb85f0e9b4a1a9d3c39600