Update patch set 2
Patch Set 2:

> Would it be easier to set the conflicts in neutron package instead
 > of -*aas packages?

re-considering it, that would be an issue for bootstraping new versions.

One collateral issue we may have with this is that we need to build new major release from neutron first, and then *aas projects. That's hard to control with DLRN so we could have FTBFS if it tries to build *aas first.

Patch-set: 2
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0